requirements.txt
Django==5.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
django-allauth==0.58.2
django-crispy-forms==2.1
django-extensions==4.1
django-heroku==0.3.1
crispy-bootstrap5==0.7
django-summernote==0.8.20.0
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
Pillow==9.5.0
python-dotenv==1.0.0

